requirements.txt
streamlit
numpy
opencv-python
pillow
pandas